test: split Kiro menu card tests
